fix typo
embadding -> embedding
