+++ seatunnel-transforms-v2/src/main/java/org/apache/seatunnel/transform/nlpmodel/embedding/remote/openai/OpenAIModel.java
@@ -15,15 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.transform.nlpmodel.embadding.remote.doubao;
+package org.apache.seatunnel.transform.nlpmodel.embedding.remote.openai;
 
+import org.apache.seatunnel.shade.com.fasterxml.jackson.core.JsonProcessingException;
 import org.apache.seatunnel.shade.com.fasterxml.jackson.core.type.TypeReference;
 import org.apache.seatunnel.shade.com.fasterxml.jackson.databind.JsonNode;
-import org.apache.seatunnel.shade.com.fasterxml.jackson.databind.node.ArrayNode;
 import org.apache.seatunnel.shade.com.fasterxml.jackson.databind.node.ObjectNode;
 import org.apache.seatunnel.shade.com.google.common.annotations.VisibleForTesting;
 
-import org.apache.seatunnel.transform.nlpmodel.embadding.remote.AbstractModel;
+import org.apache.seatunnel.transform.nlpmodel.embedding.remote.AbstractModel;
 
 import org.apache.http.client.config.RequestConfig;
 import org.apache.http.client.methods.CloseableHttpResponse;
@@ -35,17 +35,16 @@ import org.apache.http.util.EntityUtils;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
-public class DoubaoModel extends AbstractModel {
+public class OpenAIModel extends AbstractModel {
 
     private final CloseableHttpClient client;
     private final String apiKey;
     private final String model;
     private final String apiPath;
 
-    public DoubaoModel(String apiKey, String model, String apiPath, Integer vectorizedNumber) {
+    public OpenAIModel(String apiKey, String model, String apiPath, Integer vectorizedNumber) {
         super(vectorizedNumber);
         this.apiKey = apiKey;
         this.model = model;
@@ -55,6 +54,9 @@ public class DoubaoModel extends AbstractModel {
 
     @Override
     protected List<List<Float>> vector(Object[] fields) throws IOException {
+        if (fields.length > 1) {
+            throw new IllegalArgumentException("OpenAI model only supports single input");
+        }
         return vectorGeneration(fields);
     }
 
@@ -78,7 +80,7 @@ public class DoubaoModel extends AbstractModel {
         String responseStr = EntityUtils.toString(response.getEntity());
 
         if (response.getStatusLine().getStatusCode() != 200) {
-            throw new IOException("Failed to get vector from doubao, response: " + responseStr);
+            throw new IOException("Failed to get vector from openai, response: " + responseStr);
         }
 
         JsonNode data = OBJECT_MAPPER.readTree(responseStr).get("data");
@@ -97,9 +99,11 @@ public class DoubaoModel extends AbstractModel {
     }
 
     @VisibleForTesting
-    public ObjectNode createJsonNodeFromData(Object[] fields) {
-        ArrayNode arrayNode = OBJECT_MAPPER.valueToTree(Arrays.asList(fields));
-        return OBJECT_MAPPER.createObjectNode().put("model", model).set("input", arrayNode);
+    public ObjectNode createJsonNodeFromData(Object[] data) throws JsonProcessingException {
+        ObjectNode objectNode = OBJECT_MAPPER.createObjectNode();
+        objectNode.put("model", model);
+        objectNode.put("input", data[0].toString());
+        return objectNode;
     }
 
     @Override
