+++ seatunnel-transforms-v2/src/test/java/org/apache/seatunnel/transform/EmbeddingTransformFactoryTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.seatunnel.transform;
 
-import org.apache.seatunnel.transform.nlpmodel.embadding.EmbeddingTransformFactory;
+import org.apache.seatunnel.transform.nlpmodel.embedding.EmbeddingTransformFactory;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
